news/requirements.txt
amqp==5.0.5
appdirs==1.4.4
asgiref==3.3.1
billiard==3.6.3.0
black==20.8b1
celery==5.0.5
cfgv==3.2.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
distlib==0.3.1
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
identify==1.5.13
kombu==5.0.2
Markdown==3.3.3
mypy-extensions==0.4.3
nodeenv==1.5.0
pathspec==0.8.1
pre-commit==2.10.1
prompt-toolkit==3.0.16
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
vine==5.0.0
virtualenv==20.4.2
wcwidth==0.2.5
